backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# SQLAlchemy 2.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Pydantic v2
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
fastapi-users[sqlalchemy]==13.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Lunar calculation
lunar-python==1.4.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
python-dotenv==1.0.0
black==23.12.1
ruff==0.1.11
mypy==1.8.0
